Separate Hanami.boot into its own initializer
This recommendation came from @timriley -- thank you!  See https://github.com/hanami/hanami-cli/issues/438#issuecomment-5139202560

`Hanami.boot` eagerly loads/initializes all of our Hanami components.
This is great for production, so that all of the components are ready
to be used!  But for other use cases -- e.g. running a simple command
locally -- we could use the faster `Hanami.prepare` and lazily load
components when they are needed.

Eventually, it would be nice to introduce something like this to the
initializer, but it gives me test failures from a Rails model that
injects Hanami depenencies at this point in time:

```
if Rails.env.local?
  Hanami.prepare
else
  Hanami.boot
end
```

As an added bonus, since we now have our Hanami App defined at the
expected config/app.rb path, we can start using the Hanami generators
(e.g. `hanami generate action --slice library_events events.show`)
